consumers: rename consumers configs
